requirements.txt
feedparser
requests
pyyaml
beautifulsoup4
openai
